# backend/src/pennylane_support/openapi.json
{"openapi": "3.1.0", "info": {"title": "PennyLane Support API", "description": "API for PennyLane Support Platform - A community-driven support system for PennyLane coding challenges", "version": "1.0.0"}, "paths": {"/challenges/": {"get": {"tags": ["Challenges", "challenges"], "summary": "List Challenges", "description": "List all challenges with optional filtering and pagination.", "operationId": "list_challenges_challenges__get", "parameters": [{"name": "offset", "in": "query", "required": false, "schema": {"type": "integer", "default": 0, "title": "Offset"}}, {"name": "limit", "in": "query", "required": false, "schema": {"type": "integer", "maximum": 100, "default": 20, "title": "Limit"}}, {"name": "difficulty", "in": "query", "required": false, "schema": {"anyOf": [{"$ref": "#/components/schemas/ChallengeDifficulty"}, {"type": "null"}], "title": "Difficulty"}}, {"name": "category", "in": "query", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Category"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ListResponse_ChallengePublic_"}}}}, "404": {"description": "Not found"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "post": {"tags": ["Challenges", "challenges"], "summary": "Create Challenge", "description": "Create a new coding challenge.", "operationId": "create_challenge_challenges__post", "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ChallengeCreate"}}}}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ChallengePublic"}}}}, "404": {"description": "Not found"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/challenges/{challenge_id}": {"get": {"tags": ["Challenges", "challenges"], "summary": "Read Challenge", "description": "Get a single challenge by ID.", "operationId": "read_challenge_challenges__challenge_id__get", "parameters": [{"name": "challenge_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Challenge Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ChallengePublic"}}}}, "404": {"description": "Not found"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "patch": {"tags": ["Challenges", "challenges"], "summary": "Update Challenge", "description": "Update a challenge's metadata.", "operationId": "update_challenge_challenges__challenge_id__patch", "parameters": [{"name": "challenge_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Challenge Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ChallengeUpdate"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ChallengePublic"}}}}, "404": {"description": "Not found"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"tags": ["Challenges", "challenges"], "summary": "Delete Challenge", "description": "Delete a challenge.", "operationId": "delete_challenge_challenges__challenge_id__delete", "parameters": [{"name": "challenge_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Challenge Id"}}], "responses": {"204": {"description": "Successful Response"}, "404": {"description": "Not found"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/challenges/{challenge_id}/conversations": {"get": {"tags": ["Challenges", "challenges"], "summary": "Get Challenge Conversations", "description": "Get all conversations for a specific challenge with pagination.", "operationId": "get_challenge_conversations_challenges__challenge_id__conversations_get", "parameters": [{"name": "challenge_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Challenge Id"}}, {"name": "offset", "in": "query", "required": false, "schema": {"type": "integer", "default": 0, "title": "Offset"}}, {"name": "limit", "in": "query", "required": false, "schema": {"type": "integer", "maximum": 100, "default": 20, "title": "Limit"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ListResponse_ConversationPublic_"}}}}, "404": {"description": "Not found"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/conversations/": {"get": {"tags": ["Conversations", "conversations"], "summary": "List Conversations", "description": "List all support conversations with optional filtering.", "operationId": "list_conversations_conversations__get", "parameters": [{"name": "offset", "in": "query", "required": false, "schema": {"type": "integer", "default": 0, "title": "Offset"}}, {"name": "limit", "in": "query", "required": false, "schema": {"type": "integer", "maximum": 100, "default": 20, "title": "Limit"}}, {"name": "status", "in": "query", "required": false, "schema": {"anyOf": [{"$ref": "#/components/schemas/ConversationStatus"}, {"type": "null"}], "title": "Status"}}, {"name": "category", "in": "query", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Category"}}, {"name": "challenge_id", "in": "query", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Challenge Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ListResponse_ConversationPublic_"}}}}, "404": {"description": "Not found"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "post": {"tags": ["Conversations", "conversations"], "summary": "Create Conversation", "description": "Create a new support conversation with an optional initial post.", "operationId": "create_conversation_conversations__post", "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ConversationCreate"}}}}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ConversationPublic"}}}}, "404": {"description": "Not found"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/conversations/{conversation_id}": {"get": {"tags": ["Conversations", "conversations"], "summary": "Read Conversation", "description": "Get a single conversation by ID with all its posts.", "operationId": "read_conversation_conversations__conversation_id__get", "parameters": [{"name": "conversation_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Conversation Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ConversationPublic"}}}}, "404": {"description": "Not found"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "patch": {"tags": ["Conversations", "conversations"], "summary": "Update Conversation", "description": "Update a conversation's metadata.", "operationId": "update_conversation_conversations__conversation_id__patch", "parameters": [{"name": "conversation_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Conversation Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ConversationUpdate"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ConversationPublic"}}}}, "404": {"description": "Not found"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"tags": ["Conversations", "conversations"], "summary": "Delete Conversation", "description": "Delete a conversation and all its posts.", "operationId": "delete_conversation_conversations__conversation_id__delete", "parameters": [{"name": "conversation_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Conversation Id"}}], "responses": {"204": {"description": "Successful Response"}, "404": {"description": "Not found"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/conversations/{conversation_id}/posts": {"post": {"tags": ["Conversations", "conversations"], "summary": "Create Post", "description": "Add a post to an existing conversation.", "operationId": "create_post_conversations__conversation_id__posts_post", "parameters": [{"name": "conversation_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Conversation Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PostCreate"}}}}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PostPublic"}}}}, "404": {"description": "Not found"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "get": {"tags": ["Conversations", "conversations"], "summary": "List Posts", "description": "List all posts in a conversation with pagination.", "operationId": "list_posts_conversations__conversation_id__posts_get", "parameters": [{"name": "conversation_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Conversation Id"}}, {"name": "offset", "in": "query", "required": false, "schema": {"type": "integer", "default": 0, "title": "Offset"}}, {"name": "limit", "in": "query", "required": false, "schema": {"type": "integer", "maximum": 100, "default": 20, "title": "Limit"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ListResponse_PostPublic_"}}}}, "404": {"description": "Not found"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/conversations/{conversation_id}/posts/{post_id}": {"get": {"tags": ["Conversations", "conversations"], "summary": "Read Post", "description": "Get a specific post from a conversation.", "operationId": "read_post_conversations__conversation_id__posts__post_id__get", "parameters": [{"name": "conversation_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Conversation Id"}}, {"name": "post_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Post Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PostPublic"}}}}, "404": {"description": "Not found"}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/user/": {"get": {"tags": ["User", "user"], "summary": "User", "operationId": "user_user__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/User"}}}}, "404": {"description": "Not found"}}}}, "/api/health": {"get": {"tags": ["System"], "summary": "Health Check", "description": "Health check endpoint.", "operationId": "health_check_api_health_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}}, "components": {"schemas": {"ChallengeCreate": {"properties": {"challenge_id": {"type": "string", "title": "Challenge Id"}, "title": {"type": "string", "title": "Title"}, "description": {"type": "string", "title": "Description"}, "category": {"type": "string", "title": "Category"}, "difficulty": {"$ref": "#/components/schemas/ChallengeDifficulty"}, "points": {"type": "integer", "title": "Points", "default": 0}, "tags": {"items": {"type": "string"}, "type": "array", "title": "Tags"}, "learning_objectives": {"items": {"type": "string"}, "type": "array", "title": "Learning Objectives"}, "hints": {"items": {"type": "string"}, "type": "array", "title": "Hints"}}, "type": "object", "required": ["challenge_id", "title", "description", "category", "difficulty"], "title": "ChallengeCreate", "description": "Schema for creating a new challenge."}, "ChallengeDifficulty": {"type": "string", "enum": ["Beginner", "Intermediate", "Advanced"], "title": "ChallengeDifficulty"}, "ChallengePublic": {"properties": {"challenge_id": {"type": "string", "title": "Challenge Id"}, "title": {"type": "string", "title": "Title"}, "description": {"type": "string", "title": "Description"}, "category": {"type": "string", "title": "Category"}, "difficulty": {"$ref": "#/components/schemas/ChallengeDifficulty"}, "points": {"type": "integer", "title": "Points", "default": 0}, "tags": {"items": {"type": "string"}, "type": "array", "title": "Tags"}, "learning_objectives": {"items": {"type": "string"}, "type": "array", "title": "Learning Objectives"}, "hints": {"items": {"type": "string"}, "type": "array", "title": "Hints"}, "id": {"type": "integer", "title": "Id"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}, "updated_at": {"type": "string", "format": "date-time", "title": "Updated At"}}, "type": "object", "required": ["challenge_id", "title", "description", "category", "difficulty", "id", "created_at", "updated_at"], "title": "ChallengePublic", "description": "Schema for public representation of a challenge."}, "ChallengeUpdate": {"properties": {"challenge_id": {"type": "string", "title": "Challenge Id"}, "title": {"type": "string", "title": "Title"}, "description": {"type": "string", "title": "Description"}, "category": {"type": "string", "title": "Category"}, "difficulty": {"$ref": "#/components/schemas/ChallengeDifficulty"}, "points": {"type": "integer", "title": "Points", "default": 0}, "tags": {"items": {"type": "string"}, "type": "array", "title": "Tags"}, "learning_objectives": {"items": {"type": "string"}, "type": "array", "title": "Learning Objectives"}, "hints": {"items": {"type": "string"}, "type": "array", "title": "Hints"}}, "type": "object", "required": ["challenge_id", "title", "description", "category", "difficulty"], "title": "ChallengeUpdate"}, "ConversationCreate": {"properties": {"identifier": {"type": "string", "title": "Identifier"}, "challenge_id": {"type": "integer", "title": "Challenge Id"}, "topic": {"type": "string", "title": "Topic"}, "category": {"type": "string", "title": "Category"}, "status": {"$ref": "#/components/schemas/ConversationStatus", "default": "OPEN"}, "assignee": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Assignee"}}, "type": "object", "required": ["identifier", "challenge_id", "topic", "category"], "title": "ConversationCreate", "description": "Schema for creating a new conversation."}, "ConversationPublic": {"properties": {"identifier": {"type": "string", "title": "Identifier"}, "challenge_id": {"type": "integer", "title": "Challenge Id"}, "topic": {"type": "string", "title": "Topic"}, "category": {"type": "string", "title": "Category"}, "status": {"$ref": "#/components/schemas/ConversationStatus", "default": "OPEN"}, "assignee": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Assignee"}, "user": {"type": "string", "title": "User"}, "id": {"type": "integer", "title": "Id"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}, "updated_at": {"type": "string", "format": "date-time", "title": "Updated At"}, "posts": {"items": {"$ref": "#/components/schemas/PostPublic"}, "type": "array", "title": "Posts", "default": []}}, "type": "object", "required": ["identifier", "challenge_id", "topic", "category", "user", "id", "created_at", "updated_at"], "title": "ConversationPublic", "description": "Schema for public representation of a conversation."}, "ConversationStatus": {"type": "string", "enum": ["OPEN", "IN_PROGRESS", "WAITING_FOR_USER", "RESOLVED", "CLOSED"], "title": "ConversationStatus"}, "ConversationUpdate": {"properties": {"assignee": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Assignee"}, "status": {"anyOf": [{"$ref": "#/components/schemas/ConversationStatus"}, {"type": "null"}]}}, "type": "object", "title": "ConversationUpdate", "description": "Schema for updating a conversation."}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "ListResponse_ChallengePublic_": {"properties": {"items": {"items": {"$ref": "#/components/schemas/ChallengePublic"}, "type": "array", "title": "Items"}, "total": {"type": "integer", "title": "Total"}, "offset": {"type": "integer", "title": "Offset"}, "limit": {"type": "integer", "title": "Limit"}}, "type": "object", "required": ["items", "total", "offset", "limit"], "title": "ListResponse[ChallengePublic]"}, "ListResponse_ConversationPublic_": {"properties": {"items": {"items": {"$ref": "#/components/schemas/ConversationPublic"}, "type": "array", "title": "Items"}, "total": {"type": "integer", "title": "Total"}, "offset": {"type": "integer", "title": "Offset"}, "limit": {"type": "integer", "title": "Limit"}}, "type": "object", "required": ["items", "total", "offset", "limit"], "title": "ListResponse[ConversationPublic]"}, "ListResponse_PostPublic_": {"properties": {"items": {"items": {"$ref": "#/components/schemas/PostPublic"}, "type": "array", "title": "Items"}, "total": {"type": "integer", "title": "Total"}, "offset": {"type": "integer", "title": "Offset"}, "limit": {"type": "integer", "title": "Limit"}}, "type": "object", "required": ["items", "total", "offset", "limit"], "title": "ListResponse[PostPublic]"}, "PostCreate": {"properties": {"content": {"type": "string", "title": "Content"}, "conversation_id": {"type": "integer", "title": "Conversation Id"}}, "type": "object", "required": ["content"], "title": "PostCreate", "description": "Schema for creating a new post."}, "PostPublic": {"properties": {"content": {"type": "string", "title": "Content"}, "conversation_id": {"type": "integer", "title": "Conversation Id"}, "user": {"type": "string", "title": "User"}, "id": {"type": "integer", "title": "Id"}, "timestamp": {"type": "string", "format": "date-time", "title": "Timestamp"}}, "type": "object", "required": ["content", "user", "id", "timestamp"], "title": "PostPublic", "description": "Schema for public representation of a post."}, "User": {"properties": {"user_id": {"type": "string", "title": "User Id"}, "username": {"type": "string", "title": "Username"}, "email": {"type": "string", "title": "Email"}, "role": {"$ref": "#/components/schemas/UserRole"}}, "type": "object", "required": ["user_id", "username", "email", "role"], "title": "User"}, "UserRole": {"type": "string", "enum": ["support", "user"], "title": "UserRole"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}